+++ crates/validation/src/capture.rs
@@ -1,4 +1,4 @@
-use super::{indexed, reference, storage_mapping, Drivers, Error};
+use super::{indexed, reference, storage_mapping, Drivers, Error, NoOpDrivers};
 use futures::FutureExt;
 use itertools::{EitherOrBoth, Itertools};
 use proto_flow::{capture, flow};
@@ -66,14 +66,12 @@ pub async fn walk_all_captures<D: Drivers>(
                 // disable captures in response to the source system being unreachable, and we
                 // wouldn't want a validation error for a disabled task to terminate the build.
                 if capture.spec.shards.disable {
-                    let response = no_op_validation(&request);
-                    (capture, binding_models, request, Ok(response))
+                    NoOpDrivers {}.validate_capture(request.clone())
                 } else {
-                    drivers
-                        .validate_capture(request.clone())
-                        .map(|response| (capture, binding_models, request, response))
-                        .await
+                    drivers.validate_capture(request.clone())
                 }
+                .map(|response| (capture, binding_models, request, response))
+                .await
             });
 
     let validations: Vec<(
@@ -211,21 +209,6 @@ pub async fn walk_all_captures<D: Drivers>(
     }
 
     built_captures
-}
-
-// Performs a no-op validation. The result includes a mocked `resource_path` for each binding. This
-// is assumed to be valid because Flow treats resource paths as opaque, and because it never
-// compares two resource paths from different builds.
-fn no_op_validation(req: &capture::ValidateRequest) -> capture::ValidateResponse {
-    let bindings = req
-        .bindings
-        .iter()
-        .enumerate()
-        .map(|(i, _)| capture::validate_response::Binding {
-            resource_path: vec![format!("no-op-resource-path-{i}")],
-        })
-        .collect();
-    capture::ValidateResponse { bindings }
 }
 
 fn walk_capture_request<'a>(
